@@ -12,31 +12,31 @@
     "start": "concurrently \"npm run server\" \"npm run client\""
   },
   "dependencies": {
-    "@angular/common": "2.0.0-rc.5",
-    "@angular/compiler": "2.0.0-rc.5",
-    "@angular/core": "2.0.0-rc.5",
-    "@angular/forms": "0.3.0",
-    "@angular/http": "2.0.0-rc.5",
-    "@angular/platform-browser": "2.0.0-rc.5",
-    "@angular/platform-browser-dynamic": "2.0.0-rc.5",
-    "@angular/router": "3.0.0-rc.1",
-    "@ngrx/core": "1.0.0",
-    "@ngrx/store": "2.1.2",
-    "@ngrx/store-devtools": "3.0.0",
-    "@ngrx/store-log-monitor": "3.0.0",
+    "@angular/common": "^2.0.0",
+    "@angular/compiler": "^2.0.0",
+    "@angular/core": "^2.0.0",
+    "@angular/forms": "^2.0.0",
+    "@angular/http": "^2.0.0",
+    "@angular/platform-browser": "^2.0.0",
+    "@angular/platform-browser-dynamic": "^2.0.0",
+    "@angular/router": "^3.4.6",
+    "@ngrx/core": "^1.2.0",
+    "@ngrx/store": "^2.2.1",
+    "@ngrx/store-devtools": "^3.2.3",
+    "@ngrx/store-log-monitor": "^3.0.2",
     "@types/lodash": "^4.14.50",
-    "es6-promise": "3.1.2",
-    "es6-shim": "0.35.0",
-    "lodash": "4.12.0",
-    "reflect-metadata": "0.1.3",
-    "rxjs": "5.0.0-beta.6",
-    "systemjs": "0.19.27",
-    "zone.js": "0.6.12"
+    "es6-promise": "^4.0.5",
+    "es6-shim": "^0.35.3",
+    "lodash": "^4.12.0",
+    "reflect-metadata": "^0.1.9",
+    "rxjs": "^5.1.0",
+    "systemjs": "^0.20.5",
+    "zone.js": "^0.7.6"
   },
   "devDependencies": {
-    "awesome-typescript-loader": "^0.17.0",
+    "awesome-typescript-loader": "^1.1.1",
     "concurrently": "^2.0.0",
-    "core-js": "^2.4.0",
+    "core-js": "^2.4.1",
     "jasmine-core": "^2.4.1",
     "karma": "^0.13.22",
     "karma-coverage": "^1.0.0",
@@ -47,10 +47,10 @@
     "karma-webpack": "^1.7.0",
     "phantomjs-prebuilt": "^2.1.7",
     "raw-loader": "^0.5.1",
-    "source-map-loader": "^0.1.5",
-    "typescript": "^1.8.10",
-    "webpack": "^1.13.0",
-    "webpack-dev-server": "^1.14.1"
+    "source-map-loader": "^0.1.6",
+    "typescript": "^2.1.0",
+    "webpack": "^1.14.0",
+    "webpack-dev-server": "^1.16.2"
   },
   "author": "Lukas Ruebbelke",
   "license": "MIT"
