@@ -3,8 +3,24 @@ import { Widgets } from './src/widgets/widgets.component';
 import { Routes } from '@angular/router';
 
 export const routes: Routes = [
-  {path: '',            component: Items },
-  {path: 'items',      component: Items},
-  {path: 'widgets',    component: Widgets},
-  {path: '*',           component: Items }
+  {
+    path: '',
+    component: Items,
+    data: {title: 'Items'}
+  },
+  {
+    path: 'items',
+    component: Items,
+    data: {title: 'Items'}
+  },
+  {
+    path: 'widgets',
+    component: Widgets,
+    data: {title: 'Widgets'}
+  },
+  {
+    path: '*',
+    component: Items,
+    data: {title: 'Items'}
+  }
 ];
